+++ am/xsextra.h
@@ -18,8 +18,8 @@ extern "C" {
 
 uintptr_t _l1cache_op_read(int op_idx);
 uintptr_t _l1cache_data_read(int data_idx);
-void _l1cache_op_write(int op_idx, int val);
-void _l1cache_data_write(int data_idx, int val);
+void _l1cache_op_write(int op_idx, uintptr_t val);
+void _l1cache_data_write(int data_idx, uintptr_t val);
 
 // ============== Supplement L2/3Cache ==============
 
@@ -30,8 +30,8 @@ void _l3cache_cmd(int cmd);
 // ================= Supplement PMA =================
 uintptr_t _pma_get_addr(int addr_idx);
 uintptr_t _pma_get_cfg(int cfg_idx);
-void _pma_set_addr(int addr_idx, int val);
-void _pma_set_cfg(int cfg_idx, int val);
+void _pma_set_addr(int addr_idx, uintptr_t val);
+void _pma_set_cfg(int cfg_idx, uintptr_t val);
 
 // ================= Supplement MPE =================
 void _mpe_setncpu(char arg);
